requirements.txt
numpy==1.22.3
scikit-image==0.19.2
torch==1.11.0
torchvision==0.12.0
pillow==8.1.1
opencv-python==4.5.5.64
paddlepaddle
paddlehub
gradio
tensorflow==2.8.0
streamlit